+++ CHANGELOG.md
@@ -7,3 +7,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.0.0] - 2025-05-22
 ### Initial Release
 
+## [1.0.1] - 2025-05-22
+### Fixed
+- Fixing crash when poping from navigation view and count is 0 
+
+## Added
+- Adding unit test cases for navigation router